requirements.txt
robotframework==7.0
robotframework-appiumlibrary==2.0.0